+++ tests/test_launcher.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import glob
+import logging
 import os
 import shutil
 import subprocess
@@ -9,6 +10,7 @@ import sys
 import tempfile
 import time
 import unittest
+from unittest import mock
 
 from syncer import sync
 import websockets
@@ -138,6 +140,71 @@ class TestLauncher(unittest.TestCase):
         # console.log output is sent to stderr
         self.assertNotIn('DUMPIO_TEST', proc.stdout.decode())
         self.assertIn('DUMPIO_TEST', proc.stderr.decode())
+
+
+class TestLogLevel(unittest.TestCase):
+    def setUp(self):
+        self.logger = logging.getLogger('pyppeteer')
+        self.mock = mock.Mock()
+        self._orig_stderr = sys.stderr.write
+        sys.stderr.write = self.mock
+
+    def tearDown(self):
+        sys.stderr.write = self._orig_stderr
+        logging.getLogger('pyppeteer').setLevel(logging.NOTSET)
+
+    @sync
+    async def test_level_default(self):
+        browser = await launch(args=['--no-sandbox'])
+        await browser.close()
+
+        self.assertTrue(self.logger.isEnabledFor(logging.WARN))
+        self.assertFalse(self.logger.isEnabledFor(logging.INFO))
+        self.assertFalse(self.logger.isEnabledFor(logging.DEBUG))
+        self.mock.assert_not_called()
+
+    @sync
+    async def test_level_info(self):
+        browser = await launch(args=['--no-sandbox'], logLevel=logging.INFO)
+        await browser.close()
+
+        self.assertTrue(self.logger.isEnabledFor(logging.WARN))
+        self.assertTrue(self.logger.isEnabledFor(logging.INFO))
+        self.assertFalse(self.logger.isEnabledFor(logging.DEBUG))
+
+        self.assertIn('listening on', self.mock.call_args_list[0][0][0])
+
+    @sync
+    async def test_level_debug(self):
+        browser = await launch(args=['--no-sandbox'], logLevel=logging.DEBUG)
+        await browser.close()
+
+        self.assertTrue(self.logger.isEnabledFor(logging.WARN))
+        self.assertTrue(self.logger.isEnabledFor(logging.INFO))
+        self.assertTrue(self.logger.isEnabledFor(logging.DEBUG))
+
+        self.assertIn('listening on', self.mock.call_args_list[0][0][0])
+        self.assertIn('SEND', self.mock.call_args_list[2][0][0])
+        self.assertIn('RECV', self.mock.call_args_list[4][0][0])
+
+    @sync
+    async def test_connect_debug(self):
+        browser = await launch(args=['--no-sandbox'])
+        browser2 = await connect(
+            browserWSEndpoint=browser.wsEndpoint,
+            logLevel=logging.DEBUG,
+        )
+        page = await browser2.newPage()
+        await page.close()
+        await browser2.disconnect()
+        await browser.close()
+
+        self.assertTrue(self.logger.isEnabledFor(logging.WARN))
+        self.assertTrue(self.logger.isEnabledFor(logging.INFO))
+        self.assertTrue(self.logger.isEnabledFor(logging.DEBUG))
+
+        self.assertIn('SEND', self.mock.call_args_list[0][0][0])
+        self.assertIn('RECV', self.mock.call_args_list[2][0][0])
 
 
 class TestUserDataDir(unittest.TestCase):